+++ appliance/tests/test-zone-render.sh
@@ -21,6 +21,7 @@ assert_file_contains '$ORIGIN denuoweb.' "$zone_file" "zone origin"
 assert_file_contains '@ IN NS ns1.denuoweb.' "$zone_file" "NS record"
 assert_file_contains 'ns1 IN A 203.0.113.10' "$zone_file" "GLUE target A"
 assert_file_contains 'ns1 IN AAAA 2001:db8::10' "$zone_file" "GLUE target AAAA"
+assert_file_contains '_dns.ns1 IN SVCB 1 ns1 alpn=h2 dohpath=/dns-query{?dns}' "$zone_file" "RFC 9461 DoH SVCB record"
 assert_file_contains "_443._tcp IN TLSA 3 1 1 $digest" "$zone_file" "TLSA record"
 assert_file_contains '@ IN TXT "hns-dane-appliance=v0.1.8"' "$zone_file" "marker TXT"